backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
requests==2.28.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.3
